VP-596 Moved NumericRange to Theme dir
